NATIVE-556: Add ios build options
