+++ build-configs/configs/BuildConfigType.js
@@ -10,5 +10,8 @@ export type BuildConfigType = {|
   allowedHostNames: Array<string>,
   featureFlags: FeatureFlagsType,
   lightTheme: ThemeType,
-  darkTheme: ThemeType
+  darkTheme: ThemeType,
+  iosBuildOptions: {|
+    BUILD_CONFIG_APP_NAME: string
+  |}
 |}
