+++ build-configs/configs/integreat-test-cms/index.js
@@ -6,7 +6,10 @@ import type { BuildConfigType } from '../BuildConfigType'
 const IntegreatTestCmsBuildConfig: BuildConfigType = {
   ...IntegreatBuildConfig,
   cmsUrl: 'https://cms-test.integreat-app.de',
-  switchCmsUrl: 'https://cms.integreat-app.de'
+  switchCmsUrl: 'https://cms.integreat-app.de',
+  iosBuildOptions: {
+    BUILD_CONFIG_APP_NAME: 'Integreat Test Cms'
+  }
 }
 
 module.exports = IntegreatTestCmsBuildConfig
